requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
amqp==2.5.2
argparse==1.4.0
astroid==2.3.3
attrs==19.3.0
backports.functools-lru-cache==1.6.1
billiard==3.6.1.0
caniusepython3==7.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
click-log==0.3.2
click==7.0
code-annotations==0.3.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.6.0
diff-cover==2.5.2
distlib==0.3.0
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-dynamic-fixture==3.0.2
django-extensions==2.2.6
django-model-utils==3.2.0
django-rest-swagger==2.2.0
django-simple-history==2.8.0
django-waffle==0.19.0
django==1.11.27
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
edx-auth-backends==2.0.2
edx-django-release-util==0.3.5
edx-django-utils==2.0.3
edx-drf-extensions==2.4.5
edx-i18n-tools==0.5.0
edx-lint==1.4.1
edx-opaque-keys==2.0.1
edx-rest-api-client==1.9.2
factory-boy==2.12.0
faker==4.0.0
future==0.18.2
idna==2.8
importlib-metadata==1.4.0
inflect==3.0.2
isort==4.3.21
itypes==1.1.0
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.11.0
jsonfield==2.0.2
kombu==4.6.7
lazy-object-proxy==1.4.3
markupsafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==8.1.0
mysqlclient==1.4.6
newrelic==5.4.1.134
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
path.py==12.4.0           # via edx-i18n-tools
path==13.1.0              # via path.py
pbr==5.4.4
pip-tools==4.4.0
pluggy==0.13.1
polib==1.1.0              # via edx-i18n-tools
psutil==1.2.1
py==1.8.1
pycodestyle==2.5.0
pycryptodomex==3.9.4
pydocstyle==5.0.2
pygments==2.5.2           # via diff-cover
pyjwkest==1.3.2
pyjwt==1.7.1
pylint-celery==0.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pylint==2.4.2
pymongo==3.10.1
pyparsing==2.4.6
pytest-cov==2.8.1
pytest-django==3.8.0
pytest==5.3.4
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
pyyaml==5.3
redis==3.3.11
requests-oauthlib==1.3.0
requests==2.22.0
rest-condition==1.0.3
semantic-version==2.8.4
simplejson==3.17.0
six==1.14.0
slumber==0.7.1
snowballstemmer==2.0.0
social-auth-app-django==1.2.0
social-auth-core[openidconnect]==1.7.0
sqlparse==0.3.0           # via django-debug-toolbar
stevedore==1.31.0
text-unidecode==1.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
wrapt==1.11.2
zipp==1.1.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
